Refactor Transport and add tests
