+++ internal/server/auth_test.go
@@ -98,3 +98,38 @@ func TestTouchStone(t *testing.T) {
 	})
 
 }
+
+func TestPrepareConnection(t *testing.T) {
+	nineSixSix := func() time.Time { return time.Unix(1565998966, 0) }
+	sta, _ := InitState(nineSixSix)
+	pvBytes, _ := hex.DecodeString("10de5a3c4a4d04efafc3e06d1506363a72bd6d053baef123e6a9a79a0c04b547")
+	p, _ := ecdh.Unmarshal(pvBytes)
+	sta.staticPv = p.(crypto.PrivateKey)
+	sta.ProxyBook["shadowsocks"] = nil
+
+	t.Run("TLS correct", func(t *testing.T) {
+		chBytes, _ := hex.DecodeString("1603010200010001fc0303ac530b5778469dbbc3f9a83c6ac35b63aa6a70c2014026ade30f2faf0266f0242068424f320bcad49b4315a761f9f6dec32b0a403c2d8c0ab337608a694c6e411c0024130113031302c02bc02fcca9cca8c02cc030c00ac009c013c01400330039002f0035000a0100018f00000011000f00000c7777772e62696e672e636f6d00170000ff01000100000a000e000c001d00170018001901000101000b00020100002300000010000e000c02683208687474702f312e310005000501000000000033006b0069001d00204655c2c83aaed1db2e89ed17d671fcdc76dc96e36bde8840022f1bda2f31019600170041543af1f8d28b37d984073f40e8361613da502f16e4039f00656f427de0f66480b2e77e3e552e126bb0cc097168f6e5454c7f9501126a2377fb40151f6cfc007e0e002b0009080304030303020301000d0018001604030503060308040805080604010501060102030201002d00020101001c00024001001500920000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")
+		info, _, err := PrepareConnection(chBytes, sta, nil)
+		if err != nil {
+			t.Errorf("failed to get client info: %v", err)
+			return
+		}
+		if info.SessionId != 3710878841 {
+			t.Error("failed to get correct session id")
+			return
+		}
+	})
+	t.Run("TLS correct but replay", func(t *testing.T) {
+		chBytes, _ := hex.DecodeString("1603010200010001fc0303ac530b5778469dbbc3f9a83c6ac35b63aa6a70c2014026ade30f2faf0266f0242068424f320bcad49b4315a761f9f6dec32b0a403c2d8c0ab337608a694c6e411c0024130113031302c02bc02fcca9cca8c02cc030c00ac009c013c01400330039002f0035000a0100018f00000011000f00000c7777772e62696e672e636f6d00170000ff01000100000a000e000c001d00170018001901000101000b00020100002300000010000e000c02683208687474702f312e310005000501000000000033006b0069001d00204655c2c83aaed1db2e89ed17d671fcdc76dc96e36bde8840022f1bda2f31019600170041543af1f8d28b37d984073f40e8361613da502f16e4039f00656f427de0f66480b2e77e3e552e126bb0cc097168f6e5454c7f9501126a2377fb40151f6cfc007e0e002b0009080304030303020301000d0018001604030503060308040805080604010501060102030201002d00020101001c00024001001500920000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")
+		_, _, err := PrepareConnection(chBytes, sta, nil)
+		if err != nil {
+			t.Error("failed to prepare for the first time")
+			return
+		}
+		_, _, err = PrepareConnection(chBytes, sta, nil)
+		if err != ErrReplay {
+			t.Errorf("failed to return ErrReplay, got %v instead", err)
+			return
+		}
+	})
+}
